Designed the custom hook named useVpTableEffect to shorten the code of VpTable removing the useEffect from the component
